Refactoring
Move items shared by worker and wander to shared.py, require even
username-only accounts to be tuples, clean up some imports, move some
functions to utils, create new pickling functions, add some more things
to BAD_STATUSES, dump account pickle immediately if none exists, use
separate log file for wander, remove debug feature, add killed check to
coroutines while loop in launcher, clean up main.
